requirements.txt
absl-py==1.2.0
asttokens==2.0.8
astunparse==1.6.3
attrs==22.1.0
backcall==0.2.0
cachetools==5.2.0
certifi==2022.6.15.1
charset-normalizer==2.1.1
cycler==0.11.0
decorator==5.1.1
executing==1.0.0
flatbuffers==2.0.7
fonttools==4.37.1
gast==0.4.0
google-auth==2.11.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.48.1
h5py==3.7.0
htmlmin==0.1.12
idna==3.3
ImageHash==4.3.0
importlib-metadata==4.12.0
ipython==8.5.0
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
keras==2.10.0
Keras-Preprocessing==1.1.2
keras-tuner==1.1.3
kiwisolver==1.4.4
kt-legacy==1.0.4
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
missingno==0.5.1
multimethod==1.8
networkx==2.8.6
numpy==1.23.3
oauthlib==3.2.1
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.4
pandas-profiling==3.3.0
parso==0.8.3
patsy==0.5.2
pexpect==4.8.0
phik==0.12.2
pickleshare==0.7.5
Pillow==9.2.0
plotly==5.10.0
prompt-toolkit==3.0.31
protobuf==3.19.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.9.2
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.2.1
PyWavelets==1.3.0
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.1.2
scipy==1.9.1
seaborn==0.11.2
shap==0.41.0
six==1.16.0
sklearn==0.0
stack-data==0.5.0
statsmodels==0.13.2
tangled-up-in-unicode==0.2.0
tenacity==8.0.1
tensorboard==2.10.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor==2.0.1
threadpoolctl==3.1.0
tqdm==4.64.1
traitlets==5.4.0
typing-extensions==4.3.0
urllib3==1.26.12
visions==0.7.5
wcwidth==0.2.5
Werkzeug==2.2.2
wrapt==1.14.1
xgboost==1.6.2
zipp==3.8.1
